# 交付物/接口/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "财务单据智能风险审核系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/suppliers/{supplier_code}/risks": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "Supplier Risk By Code",
        "description": "按供应商编码查询风险摘要。",
        "operationId": "supplier_risk_by_code_api_v1_suppliers__supplier_code__risks_get",
        "parameters": [
          {
            "name": "supplier_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/document-versions/{document_version_id}/risk/evaluate": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Evaluate Risk",
        "description": "执行风险规则，不改变审批状态。",
        "operationId": "evaluate_risk_api_v1_document_versions__document_version_id__risk_evaluate_post",
        "parameters": [
          {
            "name": "document_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Version Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskEvaluationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/document-versions/{document_version_id}/risk/manual-reviews/{review_id}": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Update Manual Review",
        "description": "由审核人员更新复核状态，不改变审批状态。",
        "operationId": "update_manual_review_api_v1_document_versions__document_version_id__risk_manual_reviews__review_id__post",
        "parameters": [
          {
            "name": "document_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Version Id"
            }
          },
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/risk-findings": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "List Document Risk Findings",
        "description": "查询单据所有版本的风险项，供工作台和风险页使用。",
        "operationId": "list_document_risk_findings_api_v1_documents__document_id__risk_findings_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskFinding"
                  },
                  "title": "Response List Document Risk Findings Api V1 Documents  Document Id  Risk Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk-findings/{finding_id}/review-status": {
      "patch": {
        "tags": [
          "risk"
        ],
        "summary": "Patch Finding Review Status",
        "description": "按前端契约保存人工风险复核，不触碰审批状态。",
        "operationId": "patch_finding_review_status_api_v1_risk_findings__finding_id__review_status_patch",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskReviewStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskFinding"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Evidence-Input": {
        "properties": {
          "document_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          },
          "attachment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attachment Id"
          },
          "page_or_location": {
            "type": "string",
            "minLength": 1,
            "title": "Page Or Location"
          },
          "original_text": {
            "type": "string",
            "minLength": 1,
            "title": "Original Text"
          },
          "field_name": {
            "type": "string",
            "minLength": 1,
            "title": "Field Name"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Confidence"
          },
          "rule_version": {
            "type": "string",
            "minLength": 1,
            "title": "Rule Version"
          },
          "analyzed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Analyzed At"
          }
        },
        "type": "object",
        "required": [
          "attachment_id",
          "page_or_location",
          "original_text",
          "field_name",
          "confidence",
          "rule_version"
        ],
        "title": "Evidence",
        "description": "风险证据，缺一项即不能形成自动结论。"
      },
      "Evidence-Output": {
        "properties": {
          "document_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          },
          "attachment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attachment Id"
          },
          "page_or_location": {
            "type": "string",
            "minLength": 1,
            "title": "Page Or Location"
          },
          "original_text": {
            "type": "string",
            "minLength": 1,
            "title": "Original Text"
          },
          "field_name": {
            "type": "string",
            "minLength": 1,
            "title": "Field Name"
          },
          "confidence": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence"
          },
          "rule_version": {
            "type": "string",
            "minLength": 1,
            "title": "Rule Version"
          },
          "analyzed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Analyzed At"
          }
        },
        "type": "object",
        "required": [
          "attachment_id",
          "page_or_location",
          "original_text",
          "field_name",
          "confidence",
          "rule_version"
        ],
        "title": "Evidence",
        "description": "风险证据，缺一项即不能形成自动结论。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualReviewRequest": {
        "properties": {
          "reviewer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reviewer Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Evidence-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "reviewer_id",
          "status",
          "comment"
        ],
        "title": "ManualReviewRequest",
        "description": "人工复核提交内容。"
      },
      "ManualReviewResponse": {
        "properties": {
          "review_id": {
            "type": "string",
            "format": "uuid",
            "title": "Review Id"
          },
          "document_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Version Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "review_id",
          "document_version_id",
          "status"
        ],
        "title": "ManualReviewResponse",
        "description": "人工复核结果。"
      },
...
      "RiskEvaluationInput": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount"
          },
          "supplier_name": {
            "type": "string",
            "title": "Supplier Name"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Evidence-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          },
          "invoice_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Total"
          },
          "line_item_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Item Total"
          },
          "contract_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Amount"
          },
          "payment_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Amount"
          },
          "batch_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Total"
          },
          "payment_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Count"
          },
          "duplicate_account_count": {
            "type": "integer",
            "title": "Duplicate Account Count",
            "default": 0
          },
          "expense_limit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Limit"
          },
          "market_unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Unit Price"
          },
          "market_price_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Price Min"
          },
          "market_price_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Price Max"
          },
          "behavior_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Behavior Flags",
            "default": []
          },
          "supplier_risk_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supplier Risk Flags",
            "default": []
          },
          "required_attachment_count": {
            "type": "integer",
            "title": "Required Attachment Count",
            "default": 0
          },
          "attachment_count": {
            "type": "integer",
            "title": "Attachment Count",
            "default": 0
          },
          "attachment_fields_complete": {
            "type": "boolean",
            "title": "Attachment Fields Complete",
            "default": true
          },
          "duplicate_invoice": {
            "type": "boolean",
            "title": "Duplicate Invoice",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "amount",
          "supplier_name"
        ],
        "title": "RiskEvaluationInput",
        "description": "风险评估输入。"
      },
      "RiskEvaluationResponse": {
        "properties": {
          "document_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Version Id"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/RiskFinding"
            },
            "type": "array",
            "title": "Findings"
          }
        },
        "type": "object",
        "required": [
          "document_version_id",
          "findings"
        ],
        "title": "RiskEvaluationResponse",
        "description": "风险评估响应。"
      },
      "RiskFinding": {
        "properties": {
          "rule_code": {
            "type": "string",
            "title": "Rule Code"
          },
          "level": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low",
              "none"
            ],
            "title": "Level"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "matched",
              "confirmed",
              "dismissed",
              "manual_review"
            ],
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Evidence-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "actual_value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Actual Value"
          },
          "reference_value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Reference Value"
          },
          "threshold": {
            "additionalProperties": true,
            "type": "object",
            "title": "Threshold"
          },
          "suggestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestion"
          }
        },
        "type": "object",
        "required": [
          "rule_code",
          "level",
          "status",
          "message"
        ],
        "title": "RiskFinding",
        "description": "规则命中结果。"
      },
      "RiskReviewStatusRequest": {
        "properties": {
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "review_comment": {
            "type": "string",
            "title": "Review Comment",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "review_status"
        ],
        "title": "RiskReviewStatusRequest",
        "description": "风险项复核状态变更，与前端 PATCH 契约保持一致。"
      },
...
      "SupplierAnomaly": {
        "properties": {
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "document_no": {
            "type": "string",
            "title": "Document No"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "occurred_at",
          "document_no",
          "type",
          "description"
        ],
        "title": "SupplierAnomaly",
        "description": "供应商历史异常。"
      },
      "SupplierRiskResponse": {
        "properties": {
          "supplier_id": {
            "type": "string",
            "format": "uuid",
            "title": "Supplier Id"
          },
          "supplier_code": {
            "type": "string",
            "title": "Supplier Code"
          },
          "supplier_name": {
            "type": "string",
            "title": "Supplier Name"
          },
          "risk_status": {
            "type": "string",
            "title": "Risk Status",
            "default": "normal"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "blacklisted": {
            "type": "boolean",
            "title": "Blacklisted",
            "default": false
          },
          "blacklist_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blacklist Reason"
          },
          "payment_count": {
            "type": "integer",
            "title": "Payment Count",
            "default": 0
          },
          "total_paid": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Paid",
            "default": "0"
          },
          "last_payment_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Payment At"
          },
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/SupplierAnomaly"
            },
            "type": "array",
            "title": "Anomalies"
          }
        },
        "type": "object",
        "required": [
          "supplier_id",
          "supplier_code",
          "supplier_name"
        ],
        "title": "SupplierRiskResponse",
        "description": "供应商风险摘要。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}